Added support to create sets dir, updated documentation

+++ cards.sh
@@ -5,18 +5,15 @@ main () {
 	# For readability main code is wrapped as a function, and called at bottom of script. 
 	# This is so functions can be declared at the bottom of the code instead of the top 
 	
-selectedset=$2 # Second command line argument, saved as a variable for simplicity
+selectedset=$2 # Second command line argument, saved as a variable for readability
 
 if [ $# -eq 0 ]; then
-  printf -- "No arguments. Use -h for help\n" # Stops script if no command line arguments provided
+  printf -- "No arguments provided. Use -h for help\n" # Stops script if no command line arguments provided
   
-elif [ "$opt" = "?" ]
-then
-  echo "Default option executed (by default)"
-else # using getopts to parse command line options
-while getopts "l:a:n:w:r:d:hs" opt; do 
+else 
+while getopts "l:a:n:w:r:d:hs" opt; do  # using getopts to parse command line options
 	case ${opt} in
-		l ) 			# All modes simply call the appropriate fuctions, declared below
+		l ) 			# All code is seperated into fuctions, declared below
 			LEARNED
 			;;
 		a )
@@ -50,34 +47,33 @@ fi
 HELP () {
 	printf "Usage:\n"
 	printf "cards.sh <options> <flashcard set>\n\n"
-	printf -- "-l: learned mode, play only flashcards not marked as learned: cards.sh -l exampleset\n"
-	printf -- "-a: play all flashcards randomly, ignore learned list: cards.sh -a exampleset\n"
-	printf -- "-n: create a flashcard set: cards.sh -n setnamehere\n"
-	printf -- "-w: write to flashcards in a set: cards.sh -w exampleset\n"
-	printf -- "-r: reset learned cards for a set: cards.sh -r exampleset\n"
-	printf -- "-d: delete flashcard set: cards.sh -d exampleset\n"
-	printf -- "-s: List all flashcard sets in folder: cards.sh s\n"
+	printf -- "-l: learned mode, play unlearned flashcards from a set: cards.sh -l exampleset\n"
+	printf -- "-a: all mode: play both learned and unlearned from a set: cards.sh -a exampleset\n"
+	printf -- "-n: create a new flashcard set: cards.sh -n setnamehere\n"
+	printf -- "-w: write new flashcards in a set: cards.sh -w exampleset\n"
+	printf -- "-r: reset all learned cards for a set: cards.sh -r exampleset\n"
+	printf -- "-d: delete a flashcard set: cards.sh -d exampleset\n"
+	printf -- "-s: List all available flashcard sets: cards.sh s\n"
 	exit 0
 }
 
-#TODO: Write list function that is just a ls of dirs in the directory
 LIST () {	
 	ls sets/*.cards
 	exit 0
+	# Todo- prevent sets/ from appearing in front of each filename, and remove .cards extention from display as well
 } 
 
 LEARNED () {	
-	printf "Problem set selected: %s  --  Ignore learned flashcardscards mode\n" $selectedset
+	printf "Problem set selected: %s  --  Review unlearned only mode\n" $selectedset
 	
-	if [ ! -e sets/"${selectedset}.cards" ]; then # -d checks for directory of variable name. ! means code will run if false. Else, problem set directory must exist, and code continues
+	if [ ! -e sets/"${selectedset}.cards" ]; then # only runs if file does not exist
 		printf "Flashcard set does not exist\n"
 		exit 2
 	else
-		# printf "Flashcard set exists!\n\n"
 		while true
 		do
 			echo #newline
-			if [ ! -s sets/${selectedset}.cards ]; then # Simple check for empty flashcard set
+			if [ ! -s sets/${selectedset}.cards ]; then # only runs if file is blank
 			printf "Unlearned flashcard set is empty\n"
 			exit 2
 			fi
@@ -91,22 +87,17 @@ LEARNED () {
 				echo # newline
 				read -p "[a]nswer,[h]int,or [q]uit ? " choice
 				case "$choice" in
-					a|A )
-						printf "Answer: "
-						awk -F'======2======' '{print $2}' .temp.cards
-						break
-						;;
-					#* ) 
-					#	printf "Answer: " #This is to make the default behavior answer mode
-					#	awk -F'=' '{print $2}' .temp.cards
-					#	break
-					#	;;
 					h|H ) 
 						printf "Hint: "
 						awk -F'======2======' '{print $3}' .temp.cards
 						;;
 					q|Q ) printf "quit\n"
 						exit 0
+						;;
+					a|A|* )
+						printf "\nAnswer: "
+						awk -F'======2======' '{print $2}' .temp.cards
+						break
 						;;
 				esac
 			done
@@ -118,12 +109,11 @@ LEARNED () {
 					CRCT
 					;;
 				n|N )
-					# WRNG would move to unlearned pile. No need to do anything in the learned mode- just leave the card in the main unlearned file
-					printf "Card left in unlearned set\n\n"
+					printf "Card left in unlearned set\n"
 					;;
 				* ) 
 					printf "Assuming no\n"
-					printf "Card left in unlearned set\n\n"
+					printf "Card left in unlearned set\n"
 					;;
 			esac
 		done
@@ -132,14 +122,12 @@ LEARNED () {
 }
 
 ALL () {	
-	printf "Problem set selected: %s  --  Review all flashcardscards mode\n" $selectedset
+	printf "Problem set selected: %s  --  Review all fmode\n" $selectedset
 	
 	if [ ! -e sets/"${selectedset}.cards" ]; then # -d checks for directory of variable name. ! means code will run if false. Else, problem set directory must exist, and code continues
 		printf "Flashcard set does not exist\n"
 		exit 2
 	else
-		# printf "Flashcard set exists!\n\n"
-		
 		if [ ! -s sets/${selectedset}.cards ] && [ ! -s sets/${selectedset}.cards.learned ]; then # Simple check for empty flashcard set. Since this is the "all" mode, we techinically only need one to be not empty
 		printf "Flashcard set appears to be empty\n"
 		exit 2
@@ -256,9 +244,14 @@ ALL () {
 }
 
 NEWSET () {	
-	if [ -e sets/"${selectedset}.cards" ]; then # -d checks for directory of variable name. ! means code will run if false. Else, problem set directory must exist, and code continues
+	if [ -e sets/"${selectedset}.cards" ]; then 
 		printf "Flashcard set already exists\n"
 		exit 2
+	fi
+	if [ ! -d sets ]; then 
+		mkdir -p sets
+		printf "Created sets directory\n"
+		NEWSET
 	fi
 	:> sets/"$selectedset".cards
 	:> sets/"$selectedset".cards.learned
@@ -318,24 +311,41 @@ WRITETO () {
 }
 
 RESETLEARNED () {	
-	if [ ! -e sets/"${selectedset}.cards" ]; then # -d checks for directory of variable name. ! means code will run if false. Else, problem set directory must exist, and code continues
+	if [ ! -e sets/"${selectedset}.cards" ]; then 
 		printf "Flashcard set does not exist\n"
 		exit 2
 	fi
 	cat sets/${selectedset}.cards.learned >> sets/${selectedset}.cards # Append contents of .cards.learned file to bottom of .cards file
-	:> sets/${selectedset}.cards.learned # replace .cards.learned file with empty file
+	:> sets/${selectedset}.cards.learned # replace .cards.learned file with empty file of same name
 	printf "All learned cards moved to unlearned set\nexit\n"
 	exit 0
 }
 
 DELETESET () {	
-	if [ ! -e sets/"${selectedset}.cards" ]; then # -d checks for directory of variable name. ! means code will run if false. Else, problem set directory must exist, and code continues
+	if [ ! -e sets/"${selectedset}.cards" ]; then
 		printf "Flashcard set does not exist\n"
 		exit 2
 	fi
-	rm -f sets/${selectedset}.cards.learned
-	rm -f sets/${selectedset}.cards
-	printf "Flashcard set deleted\nexit\n"
+	printf "Delete flashcard set %s? [Y|n]"
+	while true; do
+				read -p "[Y|n] " choice
+				case "$choice" in
+					y|Y )
+						rm -f sets/${selectedset}.cards.learned
+						rm -f sets/${selectedset}.cards
+						printf "Flashcard set deleted\nexit\n"
+						;;
+					n|N  )
+						printf "User cancelled\n"
+						exit 0
+						;;
+					* ) 
+						printf "Invalid input. Operation cancelled\n"
+						exit 0
+						;;
+				esac
+			done
+	
 	exit 0
 }
 
@@ -343,7 +353,7 @@ CRCT () { # function to move card from the unlearned list > learned list
 	cat .temp.cards >> sets/${selectedset}.cards.learned
 	grep -Fvx -f .temp.cards sets/${selectedset}.cards > remaining.list #Writes new file without correct car in it
 	mv remaining.list sets/${selectedset}.cards #replaces cards list with the new file, missing the correct card, which is now in the .learned file
-	printf "Moved card to learned list\n\n"
+	printf "Moved card to learned list\n"
 }
 
 WRNG () { # function to move card from the learned list > unlearned list (if forgotten. Only used by review all mode)
